Created a new perfect_balancing policy for the relaxed Kdtree.
The creation of this policy has revealed a limitation on the relaxed kdtree and
led to a slight variation of the balancing algorithm. Overall trees with
dimension of 1 or 2 should be slightly less balanced; tree with dimension of 3
should be similarly balanced, trees with a dimension of 4 and above should be
slightly more balanced.